requirements.txt
flask>=2.3.0
gunicorn>=21.2.0
numpy>=1.24.0
Pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
